modify: 프로필 사진 -> 수정 페이지에서만 변경 가능하게 수정

+++ lib/features/user/views/user_profile_form_screen.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
@@ -93,7 +94,9 @@ class _UserProfileFormScreenState extends ConsumerState<UserProfileFormScreen>
 
   void _onTapNext(EditableFields field) {
     var index = EditableFields.values.indexOf(field) + 1;
-    index < _focusList.length ? _focusList[index].requestFocus() : _focusList[index].unfocus();
+    index < _focusList.length
+        ? _focusList[index].requestFocus()
+        : _focusList[index].unfocus();
   }
 
   FocusNode? _getFocusNode(EditableFields field) {
@@ -115,8 +118,51 @@ class _UserProfileFormScreenState extends ConsumerState<UserProfileFormScreen>
     }
   }
 
-  Future<void> _onTapDeleteAvatar() async {
-    await ref.read(avatarProvider.notifier).deleteAvatar(widget.profile);
+  Future<void> _onTapDeleteAvatar(UserProfileModel profile) async {
+    if (!profile.hasAvatar) return;
+
+    await _getAlert(
+      title: S.of(context).deleteProfilePicture,
+      destructiveActionCallback: _closeDialog,
+      confirmActionCallback: () async {
+        await ref.read(avatarProvider.notifier).deleteAvatar(widget.profile);
+        _closeDialog();
+      },
+    );
+  }
+
+  void _closeDialog() {
+    Navigator.of(context).pop();
+  }
+
+  Future<void> _getAlert({
+    required String title,
+    required void Function() confirmActionCallback,
+    required void Function() destructiveActionCallback,
+  }) async {
+    await showCupertinoDialog(
+        context: context,
+        builder: (context) => CupertinoAlertDialog(
+              title: Text(
+                title,
+                style: const TextStyle(
+                  fontSize: Sizes.size16,
+                  fontWeight: FontWeight.w400,
+                ),
+              ),
+              // content: const Text('Please confirm'),
+              actions: [
+                CupertinoDialogAction(
+                  onPressed: confirmActionCallback,
+                  child: const Text("Yes"),
+                ),
+                CupertinoDialogAction(
+                  onPressed: destructiveActionCallback,
+                  isDestructiveAction: true,
+                  child: const Text("No"),
+                ),
+              ],
+            ));
   }
 
   List<Widget> _getUserPic(UserProfileModel profile) {
@@ -125,25 +171,23 @@ class _UserProfileFormScreenState extends ConsumerState<UserProfileFormScreen>
       Stack(
         children: [
           Avatar(
-            isVertical: false,
             user: profile,
+            isVertical: false,
+            isEditable: true,
           ),
-          if (!ref.watch(avatarProvider).isLoading) ...[
+          if (!ref.watch(avatarProvider).isLoading && profile.hasAvatar) ...[
             Positioned(
               bottom: 0,
               right: 0,
-              child: GestureDetector(
-                onTap: _onTapDeleteAvatar,
-                child: Container(
-                  width: Sizes.size28 + Sizes.size2,
-                  height: Sizes.size28 + Sizes.size2,
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
+              child: Container(
+                width: Sizes.size28 + Sizes.size2,
+                height: Sizes.size28 + Sizes.size2,
+                decoration: BoxDecoration(
+                  shape: BoxShape.circle,
+                  color: Colors.white,
+                  border: Border.all(
                     color: Colors.white,
-                    border: Border.all(
-                      color: Colors.white,
-                      width: Sizes.size5,
-                    ),
+                    width: Sizes.size5,
                   ),
                 ),
               ),
@@ -152,7 +196,7 @@ class _UserProfileFormScreenState extends ConsumerState<UserProfileFormScreen>
               bottom: 3,
               right: 3,
               child: GestureDetector(
-                onTap: _onTapDeleteAvatar,
+                onTap: () => _onTapDeleteAvatar(profile),
                 child: FaIcon(
                   FontAwesomeIcons.solidCircleXmark,
                   color: Colors.grey.shade600,
@@ -160,7 +204,7 @@ class _UserProfileFormScreenState extends ConsumerState<UserProfileFormScreen>
                 ),
               ),
             ),
-          ]
+          ],
         ],
       ),
       Gaps.v20,
